This is a project based on Shopping Website created using Servlets.
In this project, we can add things to cart , make our account on the shopping website can login and logout. If the user is n ot login and wishes to buy the selected items the n 1st he would have to login. Ther are main groups inside which there is a subgroup.